Draw fancier materials.
We can now draw (ugly) circles.

We can also draw materials which have no color.  This isn't very
interesting other than that it represents the beginning of an API
to support different kinds of bindings for the color channel,
such as textures (not implemented in this patch).

Added support for preprocessing shader programs.  Interestingly, we
both refactored UniqueProgram at the same time.  The new version has
the ability to assemble shader programs from multiple string buffers.

Simplified the object ownership model for the scene.  Objects are
all monomorphic under the assumption that most of the heavyweight
data (such as meshes) will be referenced indirectly so we should
be able to get away with a lightweight structure in the common
case for Objects, Shapes, and Materials.

We now have a fancy polymorphic shader called MaterialShader.
MaterialShader is characterized by a MaterialShaderDescription
which is generated from properties of the Material and a Modifier.
The Modifier captures shader variations which have to do with how
drawing will take place (such as for drawing circles) as distinguished
from the Material which only describes the surface properties.